+++ packages/client/.storybook/main.js
@@ -1,13 +1,9 @@
 module.exports = {
   stories: ['../src/**/*.stories.mdx', '../src/**/*.stories.@(js|jsx|ts|tsx)'],
-  typescript: {
-    reactDocgen: 'none', // Workaround for issue https://github.com/styleguidist/react-docgen-typescript/issues/356
-  },
   addons: [
     '@storybook/preset-create-react-app',
     '@storybook/addon-essentials',
     '@storybook/addon-storysource',
     '@storybook/addon-a11y',
-    '@storybook/addon-jest',
   ],
 };