+++ packages/client/src/elements/animations/index.ts
@@ -1,5 +1,6 @@
 export { default as AnimatedLogo } from './AnimatedLogo';
 export { default as Checkmark } from './Checkmark';
+export { default as CircleLoader } from './CircleLoader';
 export { default as Heaven } from './Heaven';
 export { default as Hell } from './Hell';
 export { default as PageLoader } from './PageLoader';