+++ packages/client/src/pages/authorized/Dashboard/Dashboard.styled.ts
@@ -1,13 +1,18 @@
 import styled, { css } from 'styled-components';
 
-export const RowSection = styled.div(
-  ({ theme: { breakpoints, down } }) => css`
+interface RowSectionProps {
+  isCameraMode: boolean;
+}
+
+export const RowSection = styled.div<RowSectionProps>(
+  ({ isCameraMode, theme: { breakpoints, down } }) => css`
     display: flex;
     flex-wrap: wrap;
     gap: 100px;
-    margin: 80px 0 20px;
+    margin: ${isCameraMode ? '32px 0 20px' : '80px 0 20px'};
+    width: 100%;
 
-    ${down(breakpoints.md)} {
+    ${down(breakpoints.sm)} {
       gap: 0;
       align-items: center;
       flex-direction: column-reverse;