+++ packages/client/src/pages/authorized/Dashboard/index.tsx
@@ -1,11 +1,12 @@
 import React, { useContext, useEffect, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 
+import { CircleLoader } from '../../../elements';
 import { ConnectionState } from '../../../enums/connectionState.enum';
 import { DeviceStatus } from '../../../enums/deviceStatus.enum';
 import { useAxios } from '../../../hooks';
 import { WebSocketContext } from '../../../providers/api/WebSocketProvider/WebSocketProvider.context';
-import { registerWebPush } from '../../../utils';
+import { isCameraPreviewEnabled, registerWebPush } from '../../../utils';
 import { Title } from '../AuthorizedPages.styled';
 import { RowSection } from './Dashboard.styled';
 import CameraPreviewSection from './sections/CameraPreviewSection';
@@ -21,26 +22,33 @@ const Dashboard = () => {
     () => connectionState === ConnectionState.CONNECTED && deviceStatus === DeviceStatus.CONNECTED,
     [connectionState, deviceStatus],
   );
+  const isLoading = connectionState === ConnectionState.CONNECTING;
 
   useEffect(() => {
     void registerWebPush(axios);
   }, [axios]);
 
-  const DeviceSections = () => (
-    <>
-      <TogglingSection />
-      <CameraPreviewSection />
-    </>
-  );
+  const DeviceSections = () => {
+    if (!isConnected) {
+      return <GateDisconnected connectionState={connectionState} deviceStatus={deviceStatus} />;
+    }
+
+    return (
+      <>
+        <TogglingSection />
+        <CameraPreviewSection />
+      </>
+    );
+  };
 
   return (
     <>
       <Title data-testid="dashboard-title">{t('routes.dashboard.title')}</Title>
-      <RowSection>
-        {isConnected ? (
-          <DeviceSections />
+      <RowSection isCameraMode={isCameraPreviewEnabled()}>
+        {isLoading ? (
+          <CircleLoader variant="large" label="routes.dashboard.connecting" />
         ) : (
-          <GateDisconnected connectionState={connectionState} deviceStatus={deviceStatus} />
+          <DeviceSections />
         )}
       </RowSection>
     </>