+++ packages/client/src/pages/authorized/Settings/sections/Account/Account.styled.ts
@@ -67,7 +67,7 @@ export const IntegrationTemplateLink = styled(Link)`
 export const TokenActionsButtonsWrapper = styled.div`
   display: flex;
   gap: 10px;
-  margin: 30px 0;
+  padding: 30px 0;
 `;
 
 export const RegenerateTokenButton = styled(Button)`