+++ packages/client/src/utils/index.ts
@@ -1,7 +1,10 @@
+export { default as getApiUrl } from './getApiUrl';
 export { default as getCssColor } from './getCssColor';
 export { default as getLabelFromSource } from './getLabelFromSource';
 export { default as getPlaceholderFromSource } from './getPlaceholderFromSource';
+export { default as getPublicUrl } from './getPublicUrl';
 export { default as hexToRgba } from './hexToRgba';
 export { default as isAdmin } from './isAdmin';
+export { default as isCameraPreviewEnabled } from './isCameraPreviewEnabled';
 export { default as onlyOnDevEnv } from './onlyOnDevEnv';
 export { default as registerWebPush } from './registerWebPush';